requirements.txt
streamlit
transformers
diffusers
torch
soundfile
Pillow
numpy
torchsde
accelerate